Read a cross-reference stream's offsets as wide as the file they point into
ReadXRefStream cast the offset of every type 1 entry to an int, so in a
file past 2 GiB an object beyond that point was looked for at a wrapped,
usually negative, position, and the entry field itself was a uint, so an
offset in a field of five bytes lost its top byte before it got that far.
The field is a long now, read as wide as /W says, and the offset reaches
the cross-reference table as the long PdfReference.Position already was.
This is implementation note 21 of Appendix H, which the TODO beside it
named.

What is still an int is set down in large-file-trailer-scan.md: /Prev and
/XRefStm, read through GetInteger, and the writer's four-byte field.

+++ src/PdfPinata/Pdf.Advanced/PdfCrossReferenceStream.cs
@@ -61,7 +61,11 @@ internal sealed class PdfCrossReferenceStream : PdfTrailer  // Reference: 3.4.7 
 
         public uint Type;  // 0, 1, or 2.
 
-        public uint Field2;
+        /// <summary>
+        /// The byte offset of the object for a type 1 entry, and the number of the object stream
+        /// holding it for a type 2 one. An offset is as wide as the file, hence a long.
+        /// </summary>
+        public long Field2;
 
         public uint Field3;
